requirements.txt
fastapi
uvicorn
langchain-core
langchain==0.2.10
langchain-community==0.2.6
langchain-groq
pydantic
python-dotenv
fastembed
langchain-text-splitters
pymongo
orjson